refactor(backend): colocate elo domain into feature package
Moves the ELO controller, entities, repositories and DTOs out of the
generic controller/database/dto layer packages into
net.onelitefeather.blackhole.backend.elo, alongside the
already-colocated EloService/ChatToxicityService/EloSignalConsumer/
EloDecaySweeper/ToxicityScorer. Also fixes ReportController's imports,
the only cross-domain consumer of elo.EloReasonCode/EloTrack. Pure
physical move, no behavior change.
