+++ backend/src/main/java/net/onelitefeather/blackhole/backend/elo/EloEventRepository.java
@@ -1,10 +1,9 @@
-package net.onelitefeather.blackhole.backend.database.repository;
+package net.onelitefeather.blackhole.backend.elo;
 
 import io.micronaut.data.annotation.Repository;
 import io.micronaut.data.model.Page;
 import io.micronaut.data.model.Pageable;
 import io.micronaut.data.repository.PageableRepository;
-import net.onelitefeather.blackhole.backend.database.entities.EloEventEntity;
 
 import java.util.UUID;
