+++ backend/src/main/java/net/onelitefeather/blackhole/backend/elo/EloProfileRepository.java
@@ -1,8 +1,7 @@
-package net.onelitefeather.blackhole.backend.database.repository;
+package net.onelitefeather.blackhole.backend.elo;
 
 import io.micronaut.data.annotation.Repository;
 import io.micronaut.data.repository.PageableRepository;
-import net.onelitefeather.blackhole.backend.database.entities.EloProfileEntity;
 
 @Repository
 public interface EloProfileRepository extends PageableRepository<EloProfileEntity, String> {
